OCVN-20 Yet another header redesign in order to prevent the map from being covered and thus unclickable
