+++ ui/components/app/navigation-link/style.less
@@ -4,7 +4,7 @@ aside [role=navigation] a{
   font-size: 1.3em;
   text-transform: uppercase;
   border-bottom: 1px solid #ccc;
-  padding: 5px 0 5px 40px;
+  padding: 5px 0 5px 20px;
   text-decoration: none;
   &.active{
     background-color: #808080;
