+++ ui/components/app/style.less
@@ -1,3 +1,69 @@
 .content{
   margin-top: 1em;
 }
+
+.container-fluid{
+  padding: 0;
+}
+
+aside {
+  position: fixed!important;
+  top: 0;
+  bottom: 0;
+  height: auto;
+  background-color: #f2f2f2;
+  box-shadow: #333 1px 4px 10px;
+  @media all and (max-width: 1700px) {
+    h1 {
+      font-size: 28px;
+    }
+  }
+  @media all and (max-width: 1450px) {
+    h1 {
+      font-size: 24px;
+    }
+  }
+  h1 small {
+    display: block;
+  }
+  section.col-sm-12 {
+    padding: 15px 20px 15px 20px;
+    border-bottom: 1px solid #ccc;
+  }
+
+  section.faq {
+    text-transform: uppercase;
+    font-family: "Palatino Linotype", "Book Antiqua", Palatino, serif;
+  }
+}
+
+.years-bar {
+  position: absolute;
+  top: 0;
+  z-index: 10;
+  padding-left: 5px;
+  padding-right: 0;
+  display: table;
+  box-shadow: #333 5px 5px 10px;
+  a {
+    display: table-cell;
+    background-color: #e6e6e6;
+    color: #bbbbbb;
+    text-decoration: none;
+    padding: 5px 0 5px 20px;
+    border-right: 2px #f2f2f2 solid;
+    .glyphicon {
+      visibility: hidden;
+    }
+    &:last-child {
+      border-right: none;
+    }
+    &.active {
+      color: #333;
+      background-color: #f2f2f2;
+      .glyphicon {
+        visibility: visible;
+      }
+    }
+  }
+}
