+++ ui/components/planning/style.less
@@ -9,3 +9,8 @@
     color: white;
   }
 }
+
+.map-content{
+  padding-left: 5px;
+  overflow: hidden;
+}
